rename(ui-design-tool): rename Workspace to UIDesignCanvas
