+++ packages/ui-design-tool/components/UIDesignCanvas/types.ts
@@ -1,0 +1,7 @@
+import { CanvasArgs } from '@/api/Canvas/model';
+
+export type UIDesignCanvasProps = React.HTMLAttributes<HTMLDivElement> & {
+  initialData: CanvasArgs['children'];
+};
+
+export type UIDesignCanvasRef = HTMLDivElement;
