--- packages/ui-design-tool/components/Workspace/types.ts
@@ -1,7 +1,0 @@
-import { CanvasArgs } from '@/api/Canvas/model';
-
-export type WorkspaceProps = React.HTMLAttributes<HTMLDivElement> & {
-  initialData: CanvasArgs['children'];
-};
-
-export type WorkspaceRef = HTMLDivElement;
